#27540 Applying code review feedback: FileHashCalculatorService service exposed as an interface.
